Revert "뇌피셜실패 (for revert)"
This reverts commit 6d9c9cc520c8c46ca35db1db6559427deb79eacc.

+++ INSTAGRAM/app/src/main/java/csejeonju2019/go/kr/insta/Post.java
@@ -13,7 +13,7 @@ public class Post {
     public String author;
     public String title;
     public String body;
-    public  double datanum=NewPostActivity.datanum;
+    public  int datanum=NewPostActivity.datanum;
     public int starCount = 0;
     public List<String> stars = new ArrayList<String>();
 
@@ -21,7 +21,7 @@ public class Post {
         // Default constructor required for calls to DataSnapshot.getValue(Post.class)
     }
 
-    public Post(String uid, String author, String title, String body,double datanum) {
+    public Post(String uid, String author, String title, String body,int datanum) {
         this.uid = uid;
         this.author = author;
         this.title = title;
